WaitClean: fix infinite loop in timeout handling code
If the first wait timeouts, the sysfs_wait() will set the delay to a
negative number and return 0. This makes the loop go for second time,
but sysfs_wait() will now return 0 immediately, without even calling
select(). And so on, which effectively changes the timeout-loop into a
never ending busy-wait, which may block system shutdown.